Factor Upload/Download/Update Tests out
