+++ sdk-core/src/test/java/care/data4life/sdk/RecordServiceAdditionalResourceTypeTest.kt
@@ -510,675 +510,6 @@ class RecordServiceAdditionalResourceTypeTest : RecordServiceTestBase() {
     }
 
     @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun uploadOrDownloadData_shouldUploadData_Patient() {
-        // Given
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-        val patient = PatientBuilder.buildPatient()
-        val decryptedRecord  = DecryptedRecord(
-                null,
-                patient,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                null,
-                -1
-        )
-        decryptedRecord.identifier = RECORD_ID
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-        val uploadResult: MutableList<Pair<Attachment, List<String>>> = ArrayList()
-        val downscaledIds: MutableList<String> = ArrayList()
-        downscaledIds.add("downscaledId_1")
-        downscaledIds.add("downscaledId_2")
-        uploadResult.add(Pair<Attachment, List<String>>(patient.photo!![0], downscaledIds))
-        Mockito.`when`(
-                mockAttachmentService.uploadAttachments(
-                        ArgumentMatchers.any(),
-                        ArgumentMatchers.eq(mockAttachmentKey),
-                        ArgumentMatchers.eq(USER_ID))
-        ).thenReturn(Single.just(uploadResult))
-        Mockito.`when`(recordService.getValidHash(patient.photo!![0])).thenReturn(DATA_HASH)
-
-        // When
-        val record = recordService.uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.UPLOAD,
-                decryptedRecord,
-                null,
-                USER_ID
-        )
-
-        // Then
-        Truth.assertThat(record).isEqualTo(decryptedRecord)
-        Truth.assertThat(record.attachmentsKey).isEqualTo(mockAttachmentKey)
-        inOrder.verify(recordService).uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.UPLOAD,
-                decryptedRecord,
-                null,
-                USER_ID
-        )
-        inOrder.verify(mockCryptoService).generateGCKey()
-        inOrder.verify(mockAttachmentService).uploadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID)
-        )
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun uploadOrDownloadData_shouldDownloadData_Patient() {
-        // Given
-        val patient = PatientBuilder.buildPatient()
-        patient.photo!![0].id = "id"
-        val decryptedRecord  = DecryptedRecord(
-                RECORD_ID,
-                patient,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                mockAttachmentKey,
-                -1
-        )
-        Mockito.`when`(mockAttachmentService.downloadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID))
-        ).thenReturn(Single.just(ArrayList()))
-
-        // When
-        val record = recordService.uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.DOWNLOAD,
-                decryptedRecord,
-                null,
-                USER_ID
-        )
-
-        // Then
-        Truth.assertThat(record).isEqualTo(decryptedRecord)
-        inOrder.verify(recordService).uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.DOWNLOAD,
-                decryptedRecord,
-                null,
-                USER_ID
-        )
-        inOrder.verify(mockAttachmentService).downloadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID)
-        )
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun uploadOrDownloadData_shouldDownloadData_Medication() {
-        // Given
-        val medication = MedicationBuilder.buildMedication()
-        medication.image!![0].id = "id"
-        val decryptedRecord  = DecryptedRecord(
-                RECORD_ID,
-                medication,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                mockAttachmentKey,
-                -1
-        )
-        Mockito.`when`(mockAttachmentService.downloadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID))
-        ).thenReturn(Single.just(ArrayList()))
-
-        // When
-        val record = recordService.uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.DOWNLOAD,
-                decryptedRecord,
-                null,
-                USER_ID
-        )
-
-        // Then
-        Truth.assertThat(record).isEqualTo(decryptedRecord)
-        inOrder.verify(recordService).uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.DOWNLOAD,
-                decryptedRecord,
-                null,
-                USER_ID
-        )
-        inOrder.verify(mockAttachmentService).downloadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID)
-        )
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun uploadOrDownloadData_shouldUploadData_Medication() {
-        // Given
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-        val medication = MedicationBuilder.buildMedication()
-        val decryptedRecord  = DecryptedRecord(
-                null,
-                medication,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                null,
-                -1
-        )
-        decryptedRecord.identifier = RECORD_ID
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-        val uploadResult: MutableList<Pair<Attachment, List<String>?>> = ArrayList()
-        uploadResult.add(Pair<Attachment, List<String>?>(medication.image!![0], null))
-        Mockito.`when`(mockAttachmentService.uploadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID))
-        ).thenReturn(Single.just(uploadResult))
-        Mockito.`when`(recordService.getValidHash(medication.image!![0])).thenReturn(DATA_HASH)
-
-        // When
-        val record = recordService.uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.UPLOAD,
-                decryptedRecord,
-                null,
-                USER_ID
-        )
-
-        // Then
-        Truth.assertThat(record).isEqualTo(decryptedRecord)
-        Truth.assertThat(record.attachmentsKey).isEqualTo(mockAttachmentKey)
-        inOrder.verify(recordService).uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.UPLOAD,
-                decryptedRecord,
-                null,
-                USER_ID
-        )
-        inOrder.verify(mockCryptoService).generateGCKey()
-        inOrder.verify(mockAttachmentService).uploadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID)
-        )
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun uploadOrDownloadData_shouldUploadData_Observation() {
-        // Given
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-        val observation = ObservationBuilder.buildObservationWithComponent()
-        observation.component!![0].valueAttachment!!.id = null
-        val decryptedRecord  = DecryptedRecord(
-                null,
-                observation,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                null,
-                -1
-        )
-        decryptedRecord.identifier = RECORD_ID
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-        Mockito.`when`(recordService.getValidHash(observation.valueAttachment!!)).thenReturn(DATA_HASH)
-        val uploadResult: MutableList<Pair<Attachment?, List<String>>> = ArrayList()
-        val downscaledIds: MutableList<String> = ArrayList()
-        downscaledIds.add("downscaledId_1")
-        downscaledIds.add("downscaledId_2")
-        uploadResult.add(Pair<Attachment?, List<String>>(observation.component!![0].valueAttachment, downscaledIds))
-        uploadResult.add(Pair<Attachment?, List<String>>(observation.valueAttachment, downscaledIds))
-        Mockito.`when`(mockAttachmentService.uploadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID))
-        ).thenReturn(Single.just(uploadResult))
-        Mockito.`when`(recordService.getValidHash(observation.component!![0].valueAttachment!!)).thenReturn(DATA_HASH)
-
-        // When
-        val record = recordService.uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.UPLOAD,
-                decryptedRecord,
-                null,
-                USER_ID
-        )
-
-        // Then
-        Truth.assertThat(record).isEqualTo(decryptedRecord)
-        Truth.assertThat(record.attachmentsKey).isEqualTo(mockAttachmentKey)
-        inOrder.verify(recordService).uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.UPLOAD,
-                decryptedRecord,
-                null,
-                USER_ID
-        )
-        inOrder.verify(mockCryptoService).generateGCKey()
-        inOrder.verify(mockAttachmentService).uploadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID)
-        )
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun uploadOrDownloadData_shouldDownloadData_Observation() {
-        // Given
-        val observation = ObservationBuilder.buildObservationWithComponent()
-        observation.component!![0].valueAttachment!!.id = "id1"
-        observation.valueAttachment!!.id = "id0"
-        val decryptedRecord  = DecryptedRecord(
-                RECORD_ID,
-                observation,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                mockAttachmentKey,
-                -1
-        )
-        Mockito.`when`(mockAttachmentService.downloadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID))
-        ).thenReturn(Single.just(ArrayList()))
-
-        // When
-        val record = recordService.uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.DOWNLOAD,
-                decryptedRecord,
-                null,
-                USER_ID
-        )
-
-        // Then
-        Truth.assertThat(record).isEqualTo(decryptedRecord)
-        inOrder.verify(recordService).uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.DOWNLOAD,
-                decryptedRecord,
-                null,
-                USER_ID
-        )
-        inOrder.verify(mockAttachmentService).downloadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID)
-        )
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun uploadOrDownloadData_shouldUploadData_QuestionnaireResponse() {
-        // Given
-        val questionnaireResponse = QuestionnaireResponseBuilder.buildQuestionnaireResponse()
-        val decryptedRecord  = DecryptedRecord(
-                RECORD_ID,
-                questionnaireResponse,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                null,
-                -1
-        )
-
-        val uploadResult: MutableList<Pair<Attachment?, List<String>>> = ArrayList()
-        val downscaledIds: MutableList<String> = mutableListOf("downscaledId_1", "downscaledId_2")
-        uploadResult.add(Pair<Attachment?, List<String>>(
-                questionnaireResponse.item!![0].answer!![0].valueAttachment,
-                downscaledIds)
-        )
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-        Mockito.`when`(mockAttachmentService.uploadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID)
-        )).thenReturn(Single.just(uploadResult))
-        Mockito.`when`(recordService.getValidHash(questionnaireResponse.item!![0].answer!![0].valueAttachment!!))
-                .thenReturn(DATA_HASH)
-
-        // When
-        val record = recordService.uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.UPLOAD,
-                decryptedRecord,
-                null,
-                USER_ID
-        )
-
-        // Then
-        Truth.assertThat(record).isEqualTo(decryptedRecord)
-        Truth.assertThat(record.attachmentsKey).isEqualTo(mockAttachmentKey)
-        inOrder.verify(recordService).uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.UPLOAD,
-                decryptedRecord,
-                null,
-                USER_ID
-        )
-        inOrder.verify(mockCryptoService).generateGCKey()
-        inOrder.verify(mockAttachmentService).uploadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID)
-        )
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun uploadOrDownloadData_shouldDownloadData_QuestionnaireResponse() {
-        // Given
-        val questionnaireResponse = QuestionnaireResponseBuilder.buildQuestionnaireResponse()
-        questionnaireResponse.item!![0].answer!![0].valueAttachment!!.id = "id"
-        val decryptedRecord  = DecryptedRecord(
-                RECORD_ID,
-                questionnaireResponse,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                mockAttachmentKey,
-                -1
-        )
-        Mockito.`when`(mockAttachmentService.downloadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID))
-        ).thenReturn(Single.just(ArrayList()))
-
-        // When
-        val record = recordService.uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.DOWNLOAD,
-                decryptedRecord,
-                null,
-                USER_ID
-        )
-
-        // Then
-        Truth.assertThat(record).isEqualTo(decryptedRecord)
-        inOrder.verify(recordService).uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.DOWNLOAD,
-                decryptedRecord,
-                null,
-                USER_ID
-        )
-        inOrder.verify(mockAttachmentService).downloadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID)
-        )
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.IdUsageViolation::class,
-            DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun uploadOrDownloadData_shouldAppendAdditionalIdentifiers_Patient() {
-        //given
-        val patient = PatientBuilder.buildPatient()
-        patient.photo!!.add(AttachmentBuilder.buildAttachment(null))
-        val firstAttachment = patient.photo!![0]
-        firstAttachment.title = "image"
-        firstAttachment.hash = DATA_HASH
-        val secondAttachment = patient.photo!![1]
-        secondAttachment.title = "pdf"
-        secondAttachment.hash = DATA_HASH
-        val uploadResult: MutableList<Pair<Attachment, List<String>?>> = ArrayList()
-        uploadResult.add(Pair(firstAttachment, listOf(PREVIEW_ID, THUMBNAIL_ID)))
-        uploadResult.add(Pair<Attachment, List<String>?>(secondAttachment, null))
-        val dummyDecryptedRecord = DecryptedRecord(
-                RECORD_ID,
-                patient,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                null,
-                -1
-        )
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-        Mockito.`when`(recordService.getValidHash(patient.photo!![0])).thenReturn(DATA_HASH)
-        Mockito.`when`(recordService.getValidHash(secondAttachment)).thenReturn(DATA_HASH)
-        Mockito.`when`(mockAttachmentService.uploadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID))
-        ).thenReturn(Single.fromCallable {
-            firstAttachment.id = ATTACHMENT_ID
-            secondAttachment.id = ATTACHMENT_ID
-            uploadResult
-        })
-
-        //when
-        val pat = recordService.uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.UPLOAD,
-                dummyDecryptedRecord,
-                null,
-                USER_ID
-        ).resource
-
-        //then
-        Truth.assertThat(pat).isEqualTo(patient)
-        Truth.assertThat(pat!!.identifier).hasSize(1)
-        Truth.assertThat(pat.identifier!![0].value).isEqualTo(
-                RecordService.DOWNSCALED_ATTACHMENT_IDS_FMT + "#" + ATTACHMENT_ID + "#"
-                        + PREVIEW_ID + "#" + THUMBNAIL_ID
-        )
-        Truth.assertThat(pat.identifier!![0].assigner!!.reference).isEqualTo(PARTNER_ID)
-    }
-
-    @Test
-    @Throws(DataValidationException.IdUsageViolation::class,
-            DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun uploadOrDownloadData_shouldNotAppendAdditionalIdentifiers_Medication() {
-        //given
-        val medication = MedicationBuilder.buildMedication()
-        medication.image!!.add(AttachmentBuilder.buildAttachment(null))
-        val firstAttachment = medication.image!![0]
-        firstAttachment.title = "image"
-        val secondAttachment = medication.image!![1]
-        secondAttachment.title = "pdf"
-        val uploadResult: MutableList<Pair<Attachment, List<String>?>> = ArrayList()
-        uploadResult.add(Pair(firstAttachment, listOf(PREVIEW_ID, THUMBNAIL_ID)))
-        uploadResult.add(Pair<Attachment, List<String>?>(secondAttachment, null))
-        val dummyDecryptedRecord = DecryptedRecord(
-                RECORD_ID,
-                medication,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                null,
-                -1
-        )
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-        Mockito.`when`(recordService.getValidHash(firstAttachment)).thenReturn(DATA_HASH)
-        Mockito.`when`(recordService.getValidHash(secondAttachment)).thenReturn(DATA_HASH)
-        Mockito.`when`(mockAttachmentService.uploadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID))
-        ).thenReturn(Single.fromCallable {
-            firstAttachment.id = ATTACHMENT_ID
-            secondAttachment.id = ATTACHMENT_ID
-            uploadResult
-        })
-
-        //when
-        val med = recordService.uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.UPLOAD,
-                dummyDecryptedRecord,
-                null,
-                USER_ID
-        ).resource
-
-        //then
-        Truth.assertThat(med).isEqualTo(medication)
-    }
-
-    @Test
-    @Throws(DataValidationException.IdUsageViolation::class,
-            DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun uploadOrDownloadData_shouldAppendAdditionalIdentifiers_Observation() {
-        //given
-        val observation = ObservationBuilder.buildObservationWithComponent()
-        observation.component!![0].valueAttachment!!.id = null
-        observation.component!!.add(ObservationBuilder.buildComponent(null, AttachmentBuilder.buildAttachment(null)))
-        val firstAttachment = observation.component!![0].valueAttachment
-        firstAttachment!!.title = "image"
-        firstAttachment.hash = DATA_HASH
-        val secondAttachment = observation.component!![1].valueAttachment
-        secondAttachment!!.title = "pdf"
-        secondAttachment.hash = DATA_HASH
-        val attachment = observation.valueAttachment
-        attachment!!.title = "doc"
-        attachment.hash = DATA_HASH
-        val uploadResult: MutableList<Pair<Attachment?, List<String>?>> = ArrayList()
-        uploadResult.add(Pair<Attachment?, List<String>>(firstAttachment, listOf(PREVIEW_ID, THUMBNAIL_ID)))
-        uploadResult.add(Pair<Attachment?, List<String>?>(secondAttachment, null))
-        uploadResult.add(Pair<Attachment?, List<String>?>(attachment, null))
-        val dummyDecryptedRecord = DecryptedRecord(
-                RECORD_ID,
-                observation,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                null,
-                -1
-        )
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-        Mockito.`when`(recordService.getValidHash(firstAttachment)).thenReturn(DATA_HASH)
-        Mockito.`when`(recordService.getValidHash(secondAttachment)).thenReturn(DATA_HASH)
-        Mockito.`when`(recordService.getValidHash(attachment)).thenReturn(DATA_HASH)
-        Mockito.`when`(mockAttachmentService.uploadAttachments(
-                ArgumentMatchers.any(), ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID))
-        ).thenReturn(Single.fromCallable {
-            firstAttachment.id = ATTACHMENT_ID
-            secondAttachment.id = ATTACHMENT_ID
-            attachment.id = ATTACHMENT_ID
-            uploadResult
-        })
-
-        //when
-        val obs = recordService.uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.UPLOAD,
-                dummyDecryptedRecord,
-                null,
-                USER_ID
-        ).resource
-
-        //then
-        Truth.assertThat(obs).isEqualTo(observation)
-        Truth.assertThat(obs!!.identifier).hasSize(1)
-        Truth.assertThat(obs.identifier!![0].value).isEqualTo(
-                RecordService.DOWNSCALED_ATTACHMENT_IDS_FMT + "#" + ATTACHMENT_ID + "#" +
-                        PREVIEW_ID + "#" + THUMBNAIL_ID
-        )
-        Truth.assertThat(obs.identifier!![0].assigner!!.reference).isEqualTo(PARTNER_ID)
-    }
-
-    @Test
-    @Throws(DataValidationException.IdUsageViolation::class,
-            DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun uploadOrDownloadData_shouldAppendAdditionalIdentifiers_QuestionnaireResponse() {
-        //given
-        val questionnaireResponse = QuestionnaireResponseBuilder.buildQuestionnaireResponse()
-        questionnaireResponse.item!!.add(QuestionnaireResponseBuilder.buildItem("", AttachmentBuilder.buildAttachment(null)))
-        val firstAttachment = questionnaireResponse.item!![0].answer!![0].valueAttachment
-        firstAttachment!!.title = "image"
-        firstAttachment.hash = DATA_HASH
-        val secondAttachment = questionnaireResponse.item!![1].answer!![0].valueAttachment
-        secondAttachment!!.title = "pdf"
-        secondAttachment.hash = DATA_HASH
-        val uploadResult: MutableList<Pair<Attachment?, List<String>?>> = ArrayList()
-        uploadResult.add(Pair<Attachment?, List<String>>(firstAttachment, listOf(PREVIEW_ID, THUMBNAIL_ID)))
-        uploadResult.add(Pair<Attachment?, List<String>?>(secondAttachment, null))
-        val dummyDecryptedRecord = DecryptedRecord(
-                RECORD_ID,
-                questionnaireResponse,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                null,
-                -1
-        )
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-        Mockito.`when`(recordService.getValidHash(firstAttachment)).thenReturn(DATA_HASH)
-        Mockito.`when`(recordService.getValidHash(secondAttachment)).thenReturn(DATA_HASH)
-        Mockito.`when`(mockAttachmentService.uploadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID))
-        ).thenReturn(Single.fromCallable {
-            firstAttachment.id = ATTACHMENT_ID
-            secondAttachment.id = ATTACHMENT_ID
-            uploadResult
-        })
-
-        //when
-        val response = recordService.uploadOrDownloadData(
-                RecordService.UploadDownloadOperation.UPLOAD,
-                dummyDecryptedRecord,
-                null,
-                USER_ID
-        ).resource
-
-        //then
-        Truth.assertThat(response).isEqualTo(questionnaireResponse)
-        Truth.assertThat(response!!.identifier!!.value).isEqualTo(
-                RecordService.DOWNSCALED_ATTACHMENT_IDS_FMT + "#" + ATTACHMENT_ID + "#" +
-                        PREVIEW_ID + "#" + THUMBNAIL_ID
-        )
-        Truth.assertThat(response.identifier!!.assigner!!.reference).isEqualTo(PARTNER_ID)
-    }
-
-    @Test
     @Throws(DataValidationException.IdUsageViolation::class)
     fun cleanObsoleteAdditionalIdentifiers_shouldCleanObsoleteIdentifiers_Patient() {
         //given
