+++ sdk-core/src/test/java/care/data4life/sdk/RecordServiceTest.kt
@@ -283,739 +283,6 @@ class RecordServiceTest : RecordServiceTestBase() {
     }
 
     @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun uploadOrDownloadData_calls__uploadData() {
-        // Given
-        Mockito.doReturn(mockDecryptedRecord).`when`(recordService)._uploadData(
-                mockDecryptedRecord,
-                USER_ID
-        )
-
-        // When
-        val record = recordService.uploadOrDownloadData(RecordService.UploadDownloadOperation.UPLOAD,
-                mockDecryptedRecord,
-                null,
-                USER_ID
-        )
-
-        // Then
-        Truth.assertThat(record).isEqualTo(mockDecryptedRecord)
-        inOrder.verify(recordService)._uploadData(mockDecryptedRecord, USER_ID)
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun uploadData_calls__uploadData_when_no_new_resource_was_given() {
-        // Given
-        Mockito.doReturn(mockDecryptedRecord).`when`(recordService)._uploadData(
-                mockDecryptedRecord,
-                USER_ID
-        )
-
-        // When
-        val record = recordService.uploadData(mockDecryptedRecord, null, USER_ID)
-
-        // Then
-        Truth.assertThat(record).isEqualTo(mockDecryptedRecord)
-        inOrder.verify(recordService).uploadData(mockDecryptedRecord, null, USER_ID)
-        inOrder.verify(recordService)._uploadData(mockDecryptedRecord, USER_ID)
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun _uploadData_uploads_data() {
-        // Given
-        val document = buildDocumentReference()
-        val decryptedRecord = DecryptedRecord(
-                RECORD_ID,
-                document,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                null,
-                -1
-        )
-        val downscaledIds = listOf("downscaledId_1", "downscaledId_2")
-        val uploadResult = listOf(Pair(document.content[0].attachment, downscaledIds))
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-        Mockito.`when`(recordService.getValidHash(document.content[0].attachment)).thenReturn(DATA_HASH)
-        Mockito.`when`(mockAttachmentService.uploadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID))
-        ).thenReturn(Single.just(uploadResult))
-
-        // When
-        val record = recordService._uploadData(decryptedRecord, USER_ID)
-
-        // Then
-        Truth.assertThat(record).isEqualTo(decryptedRecord)
-        Truth.assertThat(record.attachmentsKey).isEqualTo(mockAttachmentKey)
-        inOrder.verify(recordService)._uploadData(decryptedRecord, USER_ID)
-        inOrder.verify(mockCryptoService).generateGCKey()
-        inOrder.verify(recordService).getValidHash(document.content[0].attachment)
-        inOrder.verify(mockAttachmentService).uploadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID)
-        )
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun _uploadData_throws_whenAttachmentIdIsSetDuringUploadFlow() {
-        // Given
-        val document = buildDocumentReference()
-        document.content[0].attachment.id = "unexpectedId"
-        val decryptedRecord = DecryptedRecord(
-                RECORD_ID,
-                document,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                null,
-                -1
-        )
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-
-        // When
-        try {
-            recordService._uploadData(decryptedRecord, USER_ID)
-            Assert.fail("Exception expected!")
-        } catch (e: D4LException) {
-
-            // Then
-            Truth.assertThat(e.javaClass).isEqualTo(DataValidationException.IdUsageViolation::class.java)
-            Truth.assertThat(e.message).isEqualTo("Attachment.id should be null")
-        }
-        inOrder.verify(recordService)._uploadData(decryptedRecord, USER_ID)
-        inOrder.verify(mockCryptoService).generateGCKey()
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun _uploadData_throws_whenInvalidHashAttachmentDuringUploadFlow() {
-        // Given
-        val document = buildDocumentReference()
-        document.content[0].attachment.id = null
-        document.content[0].attachment.hash = "hash"
-        val decryptedRecord = DecryptedRecord(
-                RECORD_ID,
-                document,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                null,
-                -1
-        )
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-
-        // When
-        try {
-            recordService._uploadData(decryptedRecord, USER_ID)
-            Assert.fail("Exception expected!")
-        } catch (e: D4LException) {
-
-            // Then
-            Truth.assertThat(e.javaClass).isEqualTo(DataValidationException.InvalidAttachmentPayloadHash::class.java)
-            Truth.assertThat(e.message).isEqualTo("Attachment.hash is not valid")
-        }
-        inOrder.verify(recordService)._uploadData(decryptedRecord, USER_ID)
-        inOrder.verify(mockCryptoService).generateGCKey()
-        inOrder.verify(recordService).getValidHash(document.content[0].attachment)
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.IdUsageViolation::class,
-            DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun _uploadData_appends_AdditionalIdentifiers() {
-        //given
-        val docRef = buildDocumentReference()
-        docRef.content.add(buildDocRefContent(AttachmentBuilder.buildAttachment(null)))
-        val firstAttachment = docRef.content[0].attachment
-        firstAttachment.title = "image"
-        val secondAttachment = docRef.content[1].attachment
-        secondAttachment.title = "pdf"
-        val uploadResult: MutableList<Pair<Attachment, List<String>?>> = ArrayList()
-        uploadResult.add(Pair(firstAttachment, listOf(PREVIEW_ID, THUMBNAIL_ID)))
-        uploadResult.add(Pair<Attachment, List<String>?>(secondAttachment, null))
-        val dummyDecryptedRecord = DecryptedRecord(
-                RECORD_ID,
-                docRef,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                null,
-                -1
-        )
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-        Mockito.`when`(recordService.getValidHash(secondAttachment)).thenReturn(DATA_HASH)
-        Mockito.`when`(recordService.getValidHash(firstAttachment)).thenReturn(DATA_HASH)
-        Mockito.`when`(mockAttachmentService.uploadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID))
-        ).thenReturn(Single.fromCallable {
-            firstAttachment.id = ATTACHMENT_ID
-            secondAttachment.id = ATTACHMENT_ID
-            uploadResult
-        })
-
-        //when
-        val doc = recordService._uploadData(
-                dummyDecryptedRecord,
-                USER_ID
-        ).resource
-
-        //then
-        Truth.assertThat(doc).isEqualTo(docRef)
-        Truth.assertThat(doc.identifier).hasSize(1)
-        Truth.assertThat(doc.identifier!![0].value).isEqualTo(
-                RecordService.DOWNSCALED_ATTACHMENT_IDS_FMT + "#" + ATTACHMENT_ID + "#" +
-                        PREVIEW_ID + "#" + THUMBNAIL_ID
-        )
-        Truth.assertThat(doc.identifier!![0].assigner!!.reference).isEqualTo(PARTNER_ID)
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun uploadOrDownloadData_calls_updateData() {
-        // Given
-        Mockito.doReturn(mockDecryptedRecord).`when`(recordService).updateData(
-                mockDecryptedRecord,
-                mockDocumentReference,
-                USER_ID
-        )
-
-        // When
-        val record = recordService.uploadOrDownloadData(RecordService.UploadDownloadOperation.UPDATE,
-                mockDecryptedRecord,
-                mockDocumentReference,
-                USER_ID
-        )
-
-        // Then
-        Truth.assertThat(record).isEqualTo(mockDecryptedRecord)
-        inOrder.verify(recordService).updateData(
-                mockDecryptedRecord,
-                mockDocumentReference,
-                USER_ID
-        )
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun uploadData_calls_updateData_when_a_new_resource_was_given() {
-        // Given
-        Mockito.doReturn(mockDecryptedRecord).`when`(recordService).updateData(
-                mockDecryptedRecord,
-                mockDocumentReference,
-                USER_ID
-        )
-
-        // When
-        val record = recordService.uploadData(
-                mockDecryptedRecord,
-                mockDocumentReference,
-                USER_ID
-        )
-
-        // Then
-        Truth.assertThat(record).isEqualTo(mockDecryptedRecord)
-        inOrder.verify(recordService).uploadData(
-                mockDecryptedRecord,
-                mockDocumentReference,
-                USER_ID
-        )
-        inOrder.verify(recordService).updateData(
-                mockDecryptedRecord,
-                mockDocumentReference,
-                USER_ID
-        )
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun updateData_updates_data() {
-        // Given
-        val oldDocument = buildDocumentReference()
-        val oldAttachment = oldDocument.content[0].attachment
-        oldAttachment.id = "id"
-        oldAttachment.size = 0
-        oldAttachment.hash = "hash"
-        val updatedHash = "hash2"
-        val updatedDocument = buildDocumentReference()
-        val updatedAttachment = updatedDocument.content[0].attachment
-        updatedAttachment.id = "id"
-        updatedAttachment.size = 0
-        updatedAttachment.hash = updatedHash
-        val decryptedRecord = DecryptedRecord(
-                RECORD_ID,
-                oldDocument,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                mockAttachmentKey,
-                -1
-        )
-        val downscaledIds = listOf("downscaledId_1", "downscaledId_2")
-        val uploadResult = listOf(
-                Pair(updatedDocument.content[0].attachment, downscaledIds)
-        )
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockAttachmentKey))
-        Mockito.`when`(recordService.getValidHash(updatedDocument.content[0].attachment))
-                .thenReturn(DATA_HASH)
-        Mockito.`when`(mockAttachmentService.uploadAttachments(
-                ArgumentMatchers.eq(listOf(updatedAttachment)),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID))
-        ).thenReturn(Single.just(uploadResult))
-        Mockito.`when`(recordService.getValidHash(updatedDocument.content[0].attachment)).thenReturn(updatedHash)
-
-        // When
-        recordService.updateData(decryptedRecord, updatedDocument, USER_ID)
-
-        // Then
-        inOrder.verify(recordService).updateData(decryptedRecord, updatedDocument, USER_ID)
-        inOrder.verify(recordService).getValidHash(updatedDocument.content[0].attachment)
-        inOrder.verify(mockAttachmentService).uploadAttachments(
-                ArgumentMatchers.eq(listOf(updatedAttachment)),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID)
-        )
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun updateData_throws_CoreRuntimeExceptionUnsupportedOperation_whenNewResourceIsNull() {
-        // Given
-        val document = buildDocumentReference()
-        val decryptedRecord = DecryptedRecord(
-                RECORD_ID,
-                document,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                null,
-                -1
-        )
-        try {
-            // When
-            recordService.updateData(decryptedRecord, null, USER_ID)
-        } catch (e: CoreRuntimeException) {
-            // Then
-            Truth.assertThat(e.javaClass).isEqualTo(CoreRuntimeException.UnsupportedOperation::class.java)
-        }
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class, DataValidationException.IdUsageViolation::class, DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun updateData_throws_whenAttachmentHashOrSizeNotPresent() {
-        // Given
-        val document = buildDocumentReference()
-        val decryptedRecord = DecryptedRecord(
-                RECORD_ID,
-                document,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                null,
-                -1
-        )
-
-        // When
-        try {
-            document.content[0].attachment.hash = null
-            document.content[0].attachment.size = 0
-            recordService.updateData(decryptedRecord, document, USER_ID)
-            Assert.fail("Exception expected!")
-        } catch (e: D4LException) {
-
-            // Then
-            Truth.assertThat(e.javaClass).isEqualTo(DataValidationException.ExpectedFieldViolation::class.java)
-            Truth.assertThat(e.message).isEqualTo("Attachment.hash and Attachment.size expected")
-        }
-        try {
-            document.content[0].attachment.hash = "hash"
-            document.content[0].attachment.size = null
-            recordService.updateData(decryptedRecord, document, USER_ID)
-            Assert.fail("Exception expected!")
-        } catch (e: D4LException) {
-
-            // Then
-            Truth.assertThat(e.javaClass).isEqualTo(DataValidationException.ExpectedFieldViolation::class.java)
-            Truth.assertThat(e.message).isEqualTo("Attachment.hash and Attachment.size expected")
-        }
-        inOrder.verify(recordService, Mockito.times(2)).updateData(decryptedRecord, document, USER_ID)
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun updateData_throws_whenInvalidHashAttachmentDuringUpdateFlow() {
-        // Given
-        val document = buildDocumentReference()
-        val decryptedRecord = DecryptedRecord(
-                RECORD_ID,
-                document,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                null,
-                -1
-        )
-
-        // When
-        try {
-            document.content[0].attachment.hash = "hash"
-            recordService.updateData(decryptedRecord, document, USER_ID)
-            Assert.fail("Exception expected!")
-        } catch (e: D4LException) {
-
-            // Then
-            Truth.assertThat(e.javaClass).isEqualTo(DataValidationException.InvalidAttachmentPayloadHash::class.java)
-            Truth.assertThat(e.message).isEqualTo("Attachment.hash is not valid")
-        }
-        inOrder.verify(recordService).updateData(decryptedRecord, document, USER_ID)
-        inOrder.verify(recordService).getValidHash(document.content[0].attachment)
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.IdUsageViolation::class,
-            DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun updateData_throws_whenOldAttachmentHashNotPresent() {
-        // Given
-        val oldDocument = buildDocumentReference()
-        val oldAttachment = oldDocument.content[0].attachment
-        oldAttachment.id = "id"
-        oldAttachment.size = null
-        oldAttachment.hash = null
-        val updatedDocument = buildDocumentReference()
-        val updatedAttachment = updatedDocument.content[0].attachment
-        updatedAttachment.id = "id"
-        updatedAttachment.size = 0
-        updatedAttachment.hash = "hash"
-        val decryptedRecord = DecryptedRecord(
-                RECORD_ID,
-                oldDocument,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                mockAttachmentKey,
-                -1
-        )
-        val uploadResult: MutableList<Pair<Attachment, List<String>>> = ArrayList()
-        val downscaledIds = listOf("downscaledId_1", "downscaledId_2")
-        uploadResult.add(Pair(updatedAttachment, downscaledIds))
-        Mockito.`when`(mockAttachmentService.uploadAttachments(
-                ArgumentMatchers.eq(listOf(updatedAttachment)),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID))
-        ).thenReturn(Single.just(uploadResult))
-
-        // When
-        try {
-            recordService.updateData(decryptedRecord, updatedDocument, USER_ID)
-            Assert.fail("Exception expected!")
-        } catch (e: D4LException) {
-
-            // Then
-            Truth.assertThat(e.javaClass).isEqualTo(DataValidationException.InvalidAttachmentPayloadHash::class.java)
-            Truth.assertThat(e.message).isEqualTo("Attachment.hash is not valid")
-        }
-
-
-        // Then
-        inOrder.verify(recordService).updateData(decryptedRecord, updatedDocument, USER_ID)
-        inOrder.verify(recordService).getValidHash(updatedDocument.content[0].attachment)
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.IdUsageViolation::class,
-            DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun updateData_throws_Attachment_whenHashesDontMatch() {
-        // Given
-        val oldDocument = buildDocumentReference()
-        val oldAttachment = oldDocument.content[0].attachment
-        oldAttachment.id = "id"
-        oldAttachment.size = 0
-        oldAttachment.hash = "oldHash"
-        val updatedDocument = buildDocumentReference()
-        val updatedAttachment = updatedDocument.content[0].attachment
-        updatedAttachment.id = "id"
-        updatedAttachment.size = 0
-        updatedAttachment.hash = "newHash"
-        val decryptedRecord = DecryptedRecord(
-                RECORD_ID,
-                oldDocument,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                mockAttachmentKey,
-                -1
-        )
-        val uploadResult: MutableList<Pair<Attachment, List<String>>> = ArrayList()
-        val downscaledIds = listOf("downscaledId_1", "downscaledId_2")
-        uploadResult.add(Pair(updatedAttachment, downscaledIds))
-        Mockito.`when`(mockAttachmentService.uploadAttachments(
-                ArgumentMatchers.eq(listOf(updatedAttachment)),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID))
-        ).thenReturn(Single.just(uploadResult))
-
-
-        // When
-        try {
-            recordService.updateData(decryptedRecord, updatedDocument, USER_ID)
-            Assert.fail("Exception expected!")
-        } catch (e: D4LException) {
-
-            // Then
-            Truth.assertThat(e.javaClass).isEqualTo(DataValidationException.InvalidAttachmentPayloadHash::class.java)
-            Truth.assertThat(e.message).isEqualTo("Attachment.hash is not valid")
-        }
-
-        // Then
-        inOrder.verify(recordService).updateData(decryptedRecord, updatedDocument, USER_ID)
-        inOrder.verify(recordService).getValidHash(updatedDocument.content[0].attachment)
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun updateData_throws_whenValidAttachmentIdNotPresent() {
-        // Given
-        val oldDocument = buildDocumentReference()
-        val oldAttachment = oldDocument.content[0].attachment
-        oldAttachment.id = "id1"
-        oldAttachment.size = 0
-        oldAttachment.hash = "hash"
-        val updatedDocument = buildDocumentReference()
-        val updatedAttachment = updatedDocument.content[0].attachment
-        updatedAttachment.id = "id2"
-        updatedAttachment.size = 0
-        updatedAttachment.hash = "hash"
-        val decryptedRecord = DecryptedRecord(
-                RECORD_ID,
-                oldDocument,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                null,
-                -1
-        )
-        Mockito.`when`(recordService.getValidHash(updatedDocument.content[0].attachment))
-                .thenReturn("hash")
-        // When
-        try {
-            recordService.updateData(decryptedRecord, updatedDocument, USER_ID)
-            Assert.fail("Exception expected!")
-        } catch (e: D4LException) {
-
-            // Then
-            Truth.assertThat(e.javaClass).isEqualTo(DataValidationException.IdUsageViolation::class.java)
-            Truth.assertThat(e.message).isEqualTo("Valid Attachment.id expected")
-        }
-        inOrder.verify(recordService).updateData(decryptedRecord, updatedDocument, USER_ID)
-        inOrder.verify(recordService).getValidHash(updatedDocument.content[0].attachment)
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.InvalidAttachmentPayloadHash::class,
-            DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class)
-    fun updateData_notUploadsAttachment_whenHashesMatch() {
-        // Given
-        val oldDocument = buildDocumentReference()
-        val oldAttachment = oldDocument.content[0].attachment
-        oldAttachment.id = "id"
-        oldAttachment.size = 0
-        oldAttachment.hash = "hash"
-        val updatedDocument = buildDocumentReference()
-        val updatedAttachment = updatedDocument.content[0].attachment
-        updatedAttachment.id = "id"
-        updatedAttachment.size = 0
-        updatedAttachment.hash = "hash"
-        val decryptedRecord = DecryptedRecord(
-                RECORD_ID,
-                oldDocument,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                mockAttachmentKey,
-                -1
-        )
-        Mockito.`when`(recordService.getValidHash(updatedDocument.content[0].attachment))
-                .thenReturn("hash")
-        // When
-        recordService.updateData(decryptedRecord, updatedDocument, USER_ID)
-
-        // Then
-        inOrder.verify(recordService).updateData(decryptedRecord, updatedDocument, USER_ID)
-        inOrder.verify(recordService).getValidHash(updatedDocument.content[0].attachment)
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun uploadOrDownloadData_calls_downloadData() {
-        // Given
-        Mockito.doReturn(mockDecryptedRecord).`when`(recordService).downloadData(mockDecryptedRecord, USER_ID)
-
-        // When
-        val record = recordService.uploadOrDownloadData(RecordService.UploadDownloadOperation.DOWNLOAD,
-                mockDecryptedRecord,
-                null,
-                USER_ID
-        )
-
-        // Then
-        Truth.assertThat(record).isEqualTo(mockDecryptedRecord)
-        inOrder.verify(recordService).uploadOrDownloadData(RecordService.UploadDownloadOperation.DOWNLOAD,
-                mockDecryptedRecord,
-                null,
-                USER_ID
-        )
-        inOrder.verify(recordService).downloadData(mockDecryptedRecord, USER_ID)
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.IdUsageViolation::class, DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun downloadData_shouldDownloadData() {
-        // Given
-        val document = buildDocumentReference()
-        document.content[0].attachment.id = "id"
-        val decryptedRecord = DecryptedRecord(
-                RECORD_ID,
-                document,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                mockAttachmentKey,
-                -1
-        )
-        Mockito.`when`(mockAttachmentService.downloadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID))
-        ).thenReturn(Single.just(ArrayList()))
-
-        // When
-        val record = recordService.downloadData(decryptedRecord, USER_ID)
-
-        // Then
-        Truth.assertThat(record).isEqualTo(decryptedRecord)
-        inOrder.verify(recordService).downloadData(decryptedRecord, USER_ID)
-        inOrder.verify(mockAttachmentService).downloadAttachments(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.eq(mockAttachmentKey),
-                ArgumentMatchers.eq(USER_ID)
-        )
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class)
-    fun downloadData_shouldThrow_whenAttachmentIdIsNotSetDuringDownloadFlow() {
-        // Given
-        val document = buildDocumentReference()
-        document.content[0].attachment.id = null
-        val decryptedRecord = DecryptedRecord(
-                RECORD_ID,
-                document,
-                null,
-                ArrayList(),
-                null,
-                null,
-                null,
-                null,
-                -1
-        )
-
-        // When
-        try {
-            recordService.downloadData(decryptedRecord, USER_ID)
-            Assert.fail("Exception expected!")
-        } catch (e: D4LException) {
-
-            // Then
-            Truth.assertThat(e.javaClass).isEqualTo(DataValidationException.IdUsageViolation::class.java)
-            Truth.assertThat(e.message).isEqualTo("Attachment.id expected")
-        }
-        inOrder.verify(recordService).downloadData(decryptedRecord, USER_ID)
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
     @Throws(DataValidationException.IdUsageViolation::class)
     fun cleanObsoleteAdditionalIdentifiers_shouldCleanObsoleteIdentifiers() {
         //given
